added functions for fuel and taking passenger

+++ Task4-Car/src/garage/parts/engine/api/IEngine.java
@@ -7,5 +7,5 @@ public interface IEngine {
     boolean isRunning();
     boolean start();
     boolean stop();
-    double running();
+    int running();
 }
